text-service/api/requirements.txt
aiosqlite==0.21.0
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
asyncpg==0.30.0
bcrypt==4.3.0
boto3==1.40.21
botocore==1.40.21
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
deprecation==2.1.0
dill==0.4.0
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
filelock==3.19.1
graphql-core==3.2.6
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
lia-web==0.2.3
Mako==1.3.10
mangum==0.19.0
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
passlib==1.7.4
pipenv==2025.0.4
platformdirs==4.4.0
pluggy==1.6.0
postgrest==1.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.8
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-env==1.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
realtime==2.7.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
storage3==0.12.1
strawberry-graphql==0.281.0
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
